The SyntaxWalker API doesn't handle deeply nested syntax nodes well. It's implemented using recursion, and thus susceptible to stack overflow when using deeply nested structures (such as very long binary expressions).
As this API is public, and exposes general enough functionality that it can't be implemented in a non-recursive fashion (eg, it allows overrides of the Visit method to determine the order of the work they do wrt their call to SyntaxWalker.Visit), the next best thing I can think of is to remove usage of it in roslyn.

There were only a couple usages of it, and all were easy enough to switch over to using the SyntaxNode.DescendantNodes* methods (which are non-recursive and slightly more efficient in that they create stacks of enumerators instead of stacks of nodes.