Fix race in AsyncLazy where we might not return a single result
Because our AsyncLazy supports cancellation, it does not make a
guarantee that the underlying computation will only run once. It doesn't
even strictly guarantee that once one computation completes that will
be the observed value. But it does guarantee that once any caller
of GetValue or GetValueAsync has received a value, that will be the
value returned for all remaining calls.

This however was broken in a very subtle way: if an async request was
started and then cancelled, the request can still be running if the
cancellation hasn't yet been observed. A badly timed synchronous request
could also start a synchronous computation. If the async computation
were to complete at the wrong time and ignored cancellation, we might
end up caching one value, but the synchronous computation might still
return another value.

The implementation of AsyncLazy generally tries to avoid this gotcha
by ensuring that before we ever return any value, we do a check against
the current cached result, and always return the cached result. There
was one place this pattern was missed, which was in the synchronous
GetValue where the bug lies. Ensuring that works like everything else
brings back our proper guarantee.

Fixes https://github.com/dotnet/roslyn/issues/67442