Run .NET Core MSBuildWorkspace tests on machines with VS installed
We run MSBuildWorkspace tests on our integration test queue since that
is where VS is installed, so we can have coverage for all the tests
that require a VS MSBuild to be installed. Now that we can run those
from different flavors of hosts (both .NET Framework and .NET Core)
we should now do both TFMs.