Fix exception in PackageSettingsPersister
Fixes #70740
PackageSettingsPersister attempts to get or load RoslynPackage during initialization. This might fail under certain circumstances (which are not clearly understood as of now). Instead of assuming no failure case, we start listening to package load events for the failure case and perform the initialization when RoslynPackage gets loaded in future.

Verified that fixes the repro in #70740