Call GetURI when getting a URI for a document
We had some code that slipped in that was calling
ProtocolConversions.CreateAbsoluteUri(document.FilePath), which is fine
for regular documents but skips the special handling around source
generated files, so we'd break in those cases.

This "fixes" the UriFormatExceptions people are getting, in the sense
that you won't get a UriFormatException anymore. If you try to actually
go to a source generated document, that'll still be broken until we
merge https://github.com/dotnet/vscode-csharp/pull/5858, but at least
this means something like a find reference won't completely break if
it has a location somewhere in a source generated file.