Introduce a new URI format for source generated documents
This introduces a new URI format for source-generated documents,
which includes the GUIDs of the project ID and document ID so we can
properly round trip it from the server to the client and back. This
allows a feature (like go to definition) to return this URI to VS Code,
and then VS Code can open a document with that virtual URI and send
it back to the server.

We're also choosing to put "roslyn" in the URI name just so it doesn't
conflict with any other extensions that might choose a more generic
URI name.