Standardize use of prepareMachine for ci cleanup (#71265)
Arcade has a concept of prepareMachine which is commonly (though not always) enabled alongside the -ci switch. This causes process cleanup when a build function exits. Roslyn was using this switch in some cases, but not in others, and had implemented a process cleanup with -ci was passed. In the VMR windows build, this causes roslyn to kill the host build process. Change this so that -ci no longer performs this behavior by default, and instead use -prepareMachine to have Arcade do this for us.