Run Spanish tests on CoreClr (#67748)
* Run Spanish tests on CoreClr

This extends our Spanish coverage to include both our desktop and
coreclr test runs. This closes a coverage gap that was allowing bugs
into the product.

This coverage will be CI only as it falls into our category of important
coverage but chance of getting hit on PRs to be pretty low. Hence on
balance run the check CI only.

Related to #67743

* use the hammer

* yml

* yml

* undo hammer

* Let's try this

* more

* more

* more

* experiment

* more

* Progress

* Progress

* Documentation

* Progress

* more

* More test failures

* more tests

* more

* Fix readme

* test

* PR feedback

* Fix XML doc issue

* PR feedback

---------

Co-authored-by: Jared Parsons <jared@popcornbear.org>