Separate the process of creating an event accessor parameter and the process of figuring out its type.
This allows us to perform lookup operations within the accessor without triggering the process of
determining whether the event is a WINMD event. That process could request the list of containing type
members while we are building the list, thus causing an unexpected circularity and an unexpected reentrancy
into  SourceMemberMethodSymbol.LazyMethodChecks on the same thread.

Fixes #71400.