Reduce allocations in VB's PENamedTypeSymbol
These changes bring the EnsureNestedTypesAreLoaded code in VB to be the same as that in C#. There are other optimizations in the C# implementation of this class that haven't made their way to the VB one, but this is all I want to tackle for now as I'm not (yet) very proficient in writing VB code.

I noticed this in a dmp I took against a VS session with the Roslyn sln open.