Address UIDelays in AbstractGraphProvider.GetCommands
The .Any calls in this method are showing up in prism uidelays (not particularly high though as there are only 124 expressed hits in 17.8). This PR reduces some duplicated calls in this method, reducing the number of linear searches through nodes in about half in most codepaths.

I don't have much context on this code, so I'm also looking into whether both the VB and C# derivations of this class should be getting called all the time, but any changes in that regard would be separate from this.

Partially addresses https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1936654