Do not report conflicting types errors for source-generated APIs
It is a common scenario for source generators to make `internal` attributes in an assembly via our `PostInit` callbacks. This is mostly fine, but can result in annoying CS0436 "this type conflicts" warnings. To better support this scenario, we suppress this warning for this specific case. In order to plumb this knowledge through, I did have to update a bit larger tail than I would have otherwise preferred, but we now have the building block for implementing https://github.com/dotnet/roslyn/issues/56810 as well.