Simplify how RunTests handles TFMs (#71556)
* Simplify how RunTests handles TFMs

Previously our build scripts were manually passing TFMS into RunTests to
select assemblies. That was incredibly fragile, particularly when we
were updating the TFM set in the repo. It was really easy to miss tests
this way.

Change the code to instead just say if we're requsting .NET Core, .NET
Framework or both tests. That is a much more durable ask.

* fixup

* More

* more

* More

* work around

* more

* more