Do not compile all types during ENC (#71254)
Today, when ENCing, we avoid actually emitting IL for members that were not touched by the user. However, we still calculate diagnostics for the entire compilation, which is a lot of unnecessary work. To address this, we plumb the existing symbol filter through `Symbol.ForceComplete`, avoid work on symbols that should not be completed for this ENC session.