Add IDisposable to KeyValueLogMessage
The current implementation of KeyValueMessage uses pooling with the caller needing to explicitly call Free to release the message back to the pool. The Free call appears to be commonly overlooked in usage. This PR adds the IDisposable interface to allow callers to use the "using" mechanism for disposal.

I've only changed a single caller in this PR as I wanted to verify the desirability of this change before making the effort to change more callers.