Move to 8.0 (#71507)
* Move to 8.0

This updates our TFMS to recognize that .NET 8 has shipped and .NET 6 is
no longer supported.

At the same time I am trying to update our code base to be more driven
off of `$(NetCurrent)` and `$(NetPrevious)`. This is important for
source build and VMR work as these properties are often overriden to
drive the larger build efforts. Hard coded .NET Core based TFMs
interfere with that and cause friction for those teams.

* Apply suggestions from code review

Co-authored-by: Jason Malinowski <jason@jason-m.com>

* Clean up NETSDK1206 warnings

* NETSDK1206 warnings

* Include a check for duplicate TFMs

* fix

* more

* more

* more

* more

* more

* Fix BuildBoss checks

* More

* more

* more

* more

* more

---------

Co-authored-by: Jason Malinowski <jason@jason-m.com>