Don't create the InheritanceMarginTagger for preview views.
This showed up as about 20% of the costs of the tagging occurring during lightbulb preview. The preview window doesn't have an inheritance margin, so this tagger doesn't provide any value in this context.

Note that I didn't use the TextViewRoleAttribute on the InheritanceMarginTaggerProvider class as I don't believe it has the ability to indicate that a particular role is not desired.