Simplify how we add .editorconfigs to Solution Items
We had special logic for adding .editorconfigs to Solution Items
if you're configuring severities for code fixes. The approach though
was interesting: a special service existed which would create a file
watch for the file to be created, only to then add the file to
Solution Items once the file watch got created. There were a few
problems with this:

1. The service might leak the file watcher due to races there,
   which might cause VS to crash.
2. This implicitly assumed that the .editorconfig would ultimately
   get applied. If you looked at the preview but never actually
   invoked it, we've now leaked a file watch for the rest of the process
   lifetime that will be watching for that file. Change to a different
   solution and create that file in some way? Yep, that modified your
   unrelated solution.

This just moves all the logic in the TryApplyChanges path which is
much simpler.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1922798
(the crash)