Support lexing razor comments
In order to make it easier to use the roslyn lexer in Razor, it would be helpful if roslyn's parser had limited understanding of razor trivia types. Fortunately, since `@*` is always illegal in C#, we can simply lex this as if it were a multi-line comment, add an unconditional error, and continue, so this is exactly what this change does. I've copied relevant multi-line comment tests and changed them to `@*` formats for verification purposes.