Use Regex source generator in compiler
Looking at a bug this morning and saw our Regex usages showing up in a
pref trace. It was an expected item but at the same time it's clear this
code is in a bit of a known hot spot for one of our tests. Flipping to
use the regex source generator when possible for these regex.