Ensure we don't leak the request info if a request is cancelled
Our trivial RPC system used for the BuildHost only allows a request
to be cancelled before it's sent along the wire -- once it's sent
across then it's sent for good. But if it was cancelled, we didn't clean
up the tracking for it.

This was noticed while looking at a change that was potentially adding
a cancellation token to the FlushAsync call (which is new in .NET 8.0)
and generally ensuring we are explicit with our behavior there.

Closes https://github.com/dotnet/roslyn/issues/71580