Fix bootstrap log publish
PR #71265 tried to unify machine cleanup on CI. It did not set the
`$prepareMachine` value for our bootstrap builds. That meant that
`VBCSCompiler` kept running after the build as `ExitWithExitCode` did
not kill processes and it held onto log files that blocked publishing.
That results in us being able to investigate bootstrap failures
(see #71591).

The log file is opened with `FileShare.ReadWrite` but apparently that is
not enough for publish artifacts to work