Reduce ImmutableArray allocations around AssemblyIdentity
Speedometer runs show fairly high OOP allocations inside AssemblyDataForCompilation.GetReferencedAssemblies. The use of ArrayBuilder is a bit problematic in this context, as these arrays are commonly larger than the 128 element limit, thus the entries are not getting placed back in the pool. Additionally, as the array size wasn't known up front, extra allocations are happening inside the immutable array to reach it's final size.