Change an ApplyChangeToWorkspace overload to remove some async
We had an ApplyChangeToWorkspace overload that besides being async for
acquiring the lock, also allowed async work underneath it. This can
lead to a temptation to do JTF things (like thread switches) while
holding the lock which is not a good idea; the lock is not JTF friendly
and also taking thread switchs while you hold the global lock could
easily stall solution load.

It turns out that nothing actually needed the inner thing being async:
the only two places it was async was to acquire another lock (which
would have always succeeded) or to fetch some text in an edge case,
also which would have always completed synchronously. So we can easily
change this signature.

Hat tip to @davkean for the suggestion here.