Change WriteEvent calls to WriteEventCore (#71785)
Change WriteEvent calls to WriteEventCore

WriteEvent is quite inefficient in that it takes in a params array of objects, thus allocating both the array and boxing the int/long parameters passed in. Instead, WriteEventCore should be called with an appropriately populated EventSource.EventData.

Found while investigating OOP allocations in a speedometer test.