Switch the BuildHost to build with net6.0 again
When we launch the BuildHost we generally use the dotnet.exe on the
path, which the assumption that's the one the user is using for their
development and that has the right SDKs installed. This means we can
run into some fun version issues with VS Code though, where we target
the language server on net7.0 (of this writing); in the case that
the user has only a 6.0 SDK + runtime installed globally, we'll download
a 7.0 runtime, but that's only used for the language server process.
The BuildHost will still run with the regular dotnet.exe which needs
to support 6.0 runtimes + SDKs.

This PR downgrades us to use net6.0 for just the build host. The core
change there is trivial but it did result in a few other changes:

1. The Microsoft.Build packages only support net7.0 or higher, so we
   switch back to the version that we're using for source build
   which are old enough. Since we only use them as reference assemblies,
   it doesn't really matter what we pick.
2. 6.0 builds are then missing some attributes we have on higher
   versions, so we need to start declaring them again.