Fix NUGET_PACKAGES path in pipelines
By default the NUGET_PACKAGES path has a trailing slash. In our override, we should do this as well or the msbuild property that gets generated is missing the slash, which can break some builds.

This follows the pattern in 9805678248d0fbda63f97982f37e35de71890d2f, which only applied to one of the locations that in the microbuild branch we must change.