Refactor where we do project type detection and sending telemetry
We currently had two places we looked if a project was an SDK-style
project; we had logic in the code that launches a build host, and also
inside that build host we had some older code that tried to determine
the same thing. This was worth cleaning up directly and was something
we had deferred from the earlier pull requests, but we recently got
a bug report where the older code running in the build host caused
issues when processing projects from very old SDK versions, so that
provided enough impetus to go tackle this.

Unfortunately this commit is a bit icky, but there's a few good things
all wrapped up in it:

1. The IsSdkStyleProject method in the build host is removed.
2. Since we then move the boolean from ProjectFileInfo to passing it
   around in the language server, I introduce a little record to hold
   all of the "telemetry data" that we process as a part of loading
   projects that ultimately makes it to the telemetry service. This
   cleans up a few places we were passing around tuples today.
3. Since we now need to know the preferred build host kind (which we now
   use to know if it's an SDK-style project), we refactor the
   BuildHostProcessManager to make that easier to get; I think I like
   the new methods since the fallbacks are now clearer than it was
   before.

Fixes https://github.com/dotnet/roslyn/issues/71919 (or so we hope)