Only request the necessary document and project in GraphBuilder.PopulateMapsForFileInputNode
This method previously enumerated documents and projects until it found ones with the appropriate path, creating many unused documents and projects along the way.

Instead, as the file path informtion is stored in both the projectstate and documentstate, we can walk those state objects and find the appropriate ProjectId and DocumentId. Once we have those, it's a direct shot to get the Project and Document.