Fix central package versioning picking up Version.Details.xml
When we're in a source build environment, dependencies from other
repositories that are specified in Version.Details.xml are specified
by creating a MSBuild property with the version that came from the
prior stage of the build. It's important that we still use these
properties in our Directory.Packages.props or else we'll break
source build.