CompilationOutputInfo.AssemblyPath should be a normalized path (#71975)
Part of dotnet/razor#9907

ISSUE:
If BaseIntermediateOutputPath or similar properties contain "..", then the path in CompilationOutputInfo.AssemblyPath is not canonical (contains ".."). Code that is keying off this path may end up with multiple ways to reference the same project.

FIX:
Call FileUtilities.TryNormalizeAbsolutePath on the string before returning. This is called when the project is created/has properties changed, which is infrequent, and then the value is cached.