Modify CpsDiagnosticItemSourceProvider.CreateCollectionSource to be resilient to being passed a disposed hierarchy item. (#71991)
Addresses https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1950695

This behavior regressed recently due to the fix I introduced in https://github.com/dotnet/roslyn/pull/71664. The code winds it's way down to IVsBrowseObjectContext.UnconfiguredProject via NestedHierarchyHasProjectTreeCapability => TryGetFlags. The UnconfiguredProject property throws if the hierarchy is disposed, whereas we previously would go through interop, get back a failed hresult and return false, and silently fail out of the callstack.

I've pinged several folks and haven't gotten back a response about whether sending in a disposed hierarchy item is a bug in the caller. I assume this isn't a recent change though as we were resilient to this before and probably should continue to be.