Add compilation start / stop ETW events
This adds a compilation start / stop event pair to ETW for the compiler
server. When looking at a profile of VBCSCompiler it allows us to
isolate out the time range for specific projects. This is very helpful
when tracking down compiler perf issues, particularly for issues that
only manifest when a project is built in the context of a full build vs.
compiling independently.