Support more than one frozen source generated document at once
We have special support in a Solution object to freeze a specific
source generated document to a specific text, ignoring whatever the
underlying generator might be producing at the time. This is needed
for if a user is operating from an open source generated document, since
we need to pretend the contents are whatever is currently in their text
buffer. We supported just one document simply because the only
entrypoint was GetOpenDocumentInCurrentContextWithChanges(), which only
operated on a single document -- the other source generated documents
which might be open were just left as is.

For LSP we have a model where we fork solutions to ensure that all
document contents match all open buffers. To keep that pattern working
uniformly for open source generated documents, we add here the ability
to have multiple source generated documents to be frozen. The PR here
is largely mechanical, just replacing the previous single element with
a collection, and propagating through the solution sync system.