Reduce allocations during check for duplicate analyzers
The HasDuplicates extension method always creates a HashSet of unspecified size when the number of entries being checked exceeds 2. As the CompilationWithAnalyzers Verify methods call this on large collections, there would be a significant amount of resizing during the checking for duplicates. I didn't want to change the extension method to use pooling on an unknown type parameter, and didn't want to change it to specify a capacity as I don't know the characteristics of other callers into the method.

Instead, I duplicated the HasDuplicates method directly into CompilationWithAnalyzers and modified it to use a pooled hashset (there are other uses of PooledHashSet<DiagnosticAnalyzer> in the code, so this doesn't introduce a new pool)