Reduce allocations during check for duplicate analyzers (#72144)
The existing HasDuplicates extension method always creates a HashSet of unspecified size when the number of entries being checked exceeds 2. As the CompilationWithAnalyzers Verify methods call this on large collections, there would be a significant amount of resizing during the checking for duplicates. Create a version of the extension method that uses a pooled hashset to calculate duplicates.