The _projectIdToTrackerMap member is heavily indicted in the non-devenv allocations in the CompletionTest.Completions speedometer test (over 1 GB of allocations).
As this member is private, we have the ability to tweak this to not be an ImmutableDictionary.

This PR is an attempt to see if switching to ImmutableSegmentedDictionary is advantageous (at the suggestion of Sam). It also keeps changes from an earlier PR that allows callers to CreateCompilationTrackerMap to modify the dictionary data used to create the resultant ImmutableSegmentedDictionary.

*** Speedometer profile excerpt: before changes ***