Unify our solution reading code
MSBuildWorkspace had one version that would do the reading in-process
but did support solution filters. For VS Code we have one that was
doing it out-of-proces and didn't support solution filters. This
unifies both approaches, so everybody gets out of process and solution
filter support.

The solution filter code is simplified around exception handling:
there's a bunch of places it'd call TryGetAbsolute*Path asking for it
to throw if anything went wrong. We'd then catch that exception and
instead rethrow a generic "we couldn't read it" exception that also had
no information about the underlying failure. This just simplifies all
of that so we'll just let exceptions pass through.