Make shadow copy logic a bit smarter
The shadow copy infrastructure for analyzer does a defensive shadow copy
of an analyzer every time a load was requested. There was no attempt at
de-duping these request. That means if the _same_ analyzer is passed to
10 compilations in a build, 10 shadow copies will be made. That is done
even though the CLR is only ever going to load the first one.

This change mades the shadow copy infrastructure smarter by only
creating one shadow copy per MVID. That is a reliable identifier that
can be used in these circumstance.

This reduced the number of shadow copy events when building Roslyn.sln
from 268 to 53. The total number of DLL copies this avoids is much
higher because a shadow copy event copies an analyzer and all of its
supporting resource assemblies.