Reduce ImmutableDictionary allocations under SolutionCompilationState by using ImmutableSegmentedDictionary (#72220)
The _projectIdToTrackerMap member is heavily indicted in the non-devenv allocations in the CompletionTest.Completions speedometer test (over 1 GB of allocations).

As this member is private, we have the ability to tweak this to not be an ImmutableDictionary.

This PR switches to ImmutableSegmentedDictionary. It also allows callers to CreateCompilationTrackerMap to modify the dictionary data used to create the resultant ImmutableSegmentedDictionary.