Pass full paths to make-bootstrap (#72275)
* Pass full paths to make-bootstrap

This changes the make-bootstrap command to accept a full path to the
directory where the bootstrap toolset goes vs. a name.

Commands that call make-bootstrap need to know what the full path of the
bootstrap toolset is cause it is how we specify it during build. The old
way meant that all callers had to _know_ what the full dir was based on
a name. That spread the logic of where the make-bootstrap script put the
toolset into every caller. Inefficent.

Further locally it made debugging bootstrap toolsets more difficult than
it needed to be. Much easier to move it outside of _artifacts_ so that
we can easily delet _artifacts_ between profiling runs.

* inline the variable

* typo

* Apply suggestions from code review

Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>

---------

Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>