Address timing issue recently in CreateCompilationTrackerMap (#72350)
This was expressing itself as instability in unit tests. The underlying issue (that Sam found!) was that the CreateCompilationTrackerMap was recently modified such that it dereferenced _projectIdToTrackerMap several times. Unfortunately, this value can be changed during this method's execution, so there was no guarantee that all dereferences were using the same object. The primary fix here is to use a local that guarantees they all use the same object.

Without this change, I would hit several failures in TestCommitAfterEnterOnBlankLine when running the test locally several thousand times. With this change, I no longer hit a failure when running the test locally 10K times.

There are a couple other small changes that Sam suggested as part of this PR, primarily being a bit more precise in the callback's return value indicating whether the trackermap was modified.