Delay create the SkeletonReferenceCache
The only method which populates the object with a version and skeleton reference set is TryGetOrCreateReferenceSetAsync. The only way to get to that method is through GetOrBuildReferenceAsync.

The change in this PR is to make the caller to GetOrBuildReferenceAsync ensure that the SkeletonReferenceCache is created, all other code paths instead access the SkeletonReferenceCache via TryGetSkeletonReferenceCache and gracefully handle null.

When opening Roslyn, I was seeing around 600K of these objects are created before my change, and 0 after my change. This was noticed in inspecting a profile from the RichCopyTest speedometer test, where SkeletonReferenceCache.Clone was accounting for about 0.5% of allocations.