Reduce allocations required to create a TransformNode
Instead of wrapping the user function and passing that into the TransformNode ctor, we instead give the option of passing in a flag to indicate whether the TransformNode should do the try/catch itself.

I believe this will save about 0.5% in allocations in the RichCopyTest.CopyPlain speedometer test and will do a test insertion to validate, keeping this as a draft PR until validated.