WIP (try2): Attempt at reducing the file I/O under the AnalyzerAssemblyLoader folder.
Assemblies and their corresponding resource dlls are copied/deleted under this folder on solution open. When opening Rosly, I see about 700 dlls copied/deleted under this folder. Over 90% of these dlls are resource dlls, not something in use for my setup.

This change is very rough, but attempts to separate the code that ensures the assembly is shadow-copied properly, from the code that ensures that it's supporting resource assemblies are shadow-copied properly.

With this change locally, I see over 90% reduction in the number of these file I/O operations.