Reduce closure allocations associated with AsyncLazy usage
This was noticed while looking at speedometer profiles. This class is used in several high traffic areas, and the current AsyncLazy design doesn't allow usage in a way that prevents closure allocations. This PR adds the "arg" usage pattern to AsyncLazy such that funcs passed to AsyncLazy can take advantage of the arg infrastructure to avoid closure allocations.

This is done by changing AsyncLazy<T> to be an abstract class with two derivations:

1) AsyncLazyWithoutArg
2) AsyncLazyWithArg

The former is the same behavior as the prior implementation of AsyncLazy, holding onto the same compute functions as before. The latter holds onto an extra piece of data and compute functions which take that additional argument.

I did talk with Cyrus about an alternate approach where the base class held on to data and there was a derived type which masked that for users that didn't wish to use that mechanism. That approach had the advantage of not having the virtual calls that the approach outlined in this PR has, however, the type declarations were a bit clunky in that they needed to specify the data type. Anothed disadvantage of that approach was that all users of AsyncLazy would pay the cost of having some additional data, even if they weren't using that feature.

The first commit in this PR is the guts of the change, pretty much solely contained within the AsyncLazy`1.cs file. The second commit will contain the changes for the usages I found that were allocating closures.