Reduce closure allocations associated with AsyncLazy usage (#72449)
 Reduce closure allocations associated with AsyncLazy usage

This was noticed while looking at speedometer profiles. This class is used in several high traffic areas, and the current AsyncLazy design doesn't allow usage in a way that prevents closure allocations. This PR adds the "arg" usage pattern to AsyncLazy such that funcs passed to AsyncLazy can take advantage of the arg infrastructure to avoid closure allocations.

This is done by changing AsyncLazy<T> to be an abstract class with a single derivation that caches the argument data. Users still declare their type to be AsyncLazy<T>, but the AsyncLazy.Create methods have been altered to have optional overloads that take in data.