Better BlobBuilder pooling (#72383)
Looking at profiles building Compilers.slnf locally found a number of places that we weren't using PooledBlobBuilder that were impacting allocations. This change removes 250MB of byte[] allocations and saves 50MB on LOH.

This reduces time spent in GC during build by ~1 second.

Note: profiling the compiler server is imprecise as due to the parallel nature of the server and msbuild profiles can be quite noisy. For these measurements I did several measurements and picked the most average for the comparisons.