Reduce File I/O under the AnalyzerAssemblyLoader folder (#72412)
Assemblies and their corresponding resource dlls are copied/deleted under this folder on solution open. When opening Rosly, I see about 700 dlls copied/deleted under this folder. Over 90% of these dlls are resource dlls, not something in use for my setup.

This change separates the code that ensures the assembly is shadow-copied properly from the code that ensures that it's supporting resource assemblies are shadow-copied properly. Resource assemblies as copied on assembly load demand.

With this change locally, I see over 90% reduction in the number of these file I/O operations.