Replay Binary Log Tool
This adds a tool for replaying compilation events from binary logs
directly to the compiler server. This is an effective tool for profiling
the compiler server as it removes MSBuild overhead from the logs leaving
just the compiler.

Further because it's integrated into our code base it makes it very easy
to evaluate the efficacy of performance fixes. One simply needs to run
`replay` with and without the changes against the same binary log and
evaluate the differences in perfview.