Pass the temp path to DesktopStrongNameProvider ctor (#72587)
This addresses feedback ticket: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1950505

The underlying issue here is that the DesktopStrongNameProvider assumes it's caller will pass it a temp directory as EmitStream.CreateStream fails if the path isn't set. Jared indicated the compiler has a ban on calling Path.GetTempPath within their codebase, so it's on the caller to specify a value.